requirements.txt
alembic==1.13.0
asyncpg==0.29.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
vk_api==11.9.9
werkzeug==3.0.1
psycopg2-binary==2.9.9